Wait & Reset
Now if will notify you, then wait 30 minutes. at any point if the status of your computer changes from on state to another (under 20 to between to over 80), it resets the timer. After the 30 minutes, will send another notification if required, but new notification will only last 5 seconds so as to not be annoying. Will then repeat that over and over again until status changes.